+++ src/lib/provable/test/runtime-table.unit-test.ts
@@ -5,7 +5,8 @@ import { Random } from '../../testing/property.js';
 import { assert } from '../gadgets/common.js';
 import { RuntimeTable } from '../gadgets/runtime-table.js';
 import { constraintSystem, contains } from '../../testing/constraint-system.js';
-import { FeatureFlags, Cache } from 'o1js';
+import { FeatureFlags } from '../../proof-system/feature-flags.js';
+import { Cache } from '../../proof-system/cache.js';
 
 let uint = (n: number | bigint): Spec<bigint, Field> => {
   return fieldWithRng(Random.bignat((1n << BigInt(n)) - 1n));